remove mantle
