+++ Package.swift
@@ -7,8 +7,7 @@ let package = Package(
     name: "Molusca",
     products: [
         .executable(name: "Molusca", targets: ["Molusca"]),
-        .library(name: "XcodeTemplate", targets: ["XcodeTemplate"]),
-        .library(name: "Mantle", targets: ["Mantle"])
+        .library(name: "XcodeTemplate", targets: ["XcodeTemplate"])
     ],
     dependencies: [
         .package(url: "https://github.com/wendyliga/ConsoleIO.git", from: "1.0.0"),
@@ -17,9 +16,8 @@ let package = Package(
         .package(url: "https://github.com/apple/swift-argument-parser", .upToNextMinor(from: "0.0.1")),
     ],
     targets: [
-        .target(name: "Molusca", dependencies: ["SwiftKit", "XcodeTemplate", "Explorer", "ConsoleIO", "ArgumentParser", "Mantle"]),
+        .target(name: "Molusca", dependencies: ["SwiftKit", "XcodeTemplate", "Explorer", "ConsoleIO", "ArgumentParser"]),
         .target(name: "XcodeTemplate", dependencies: []),
-        .target(name: "Mantle", dependencies: ["SwiftKit", "XcodeTemplate", "Explorer", "ConsoleIO"]),
         .testTarget(name: "MoluscaTests", dependencies: ["Molusca"]),
         .testTarget(name: "XcodeTemplateTests", dependencies: ["XcodeTemplate"]),
     ]
